homework is done

+++ Homework/Tutorial_1/Task_5.py
@@ -8,12 +8,12 @@
 import math
 
 print('Введите коордиту А: ')
-x = float(input('x = '))
-y = float(input('y = '))
+x = int(input('x = '))
+y = int(input('y = '))
 
 print('Введите коордиту B: ')
-x1 = float(input('x = '))
-y1 = float(input('y = '))
+x1 = int(input('x = '))
+y1 = int(input('y = '))
 
 result = float('%.2f' % (math.sqrt((x - x1)**2 + (y - y1)**2)))
 print(result)
